Modify y-axis

Customize the scale limits and breaks for a y-axis that maps a continuous variable.

jp_dw_scale_y(chart_id, y_var, max = "max", min = "min", num_ticks = 5)

Arguments

chart_id
String. Datawrapper chart id.

y_var
String. The variable name mapped on the y-axis.

max
Integer. The upper limit of the scale. Defaults to the highest value in the variable.

min
The lower limit of the scale. Defaults to the lowest value in the variable. Also takes an integer and “truncated”, which leaves some one-third space of the plot area at the bottom empty for a chart that doesn’t start at zero.

num_ticks
Integer. Number of axis ticks. Defaults to 5. The total number of ticks that appears on the chart will be num_ticks + 1 as it includes the lower limit of the axis.

See also

Examples

if (FALSE) {
jp_dw_scale_y(
  chart_id = "4BcD3",
  y_var = "lifeExp",
  min = 0,
  num_ticks = 5
)
}