Customize y-axis of Datawrapper charts

Customize the y-axis limits and major breaks of Datawrapper charts.

jp_scale_y(
  chart_id,
  y,
  scale_y_max = max(y),
  scale_y_min_rule = "min",
  scale_y_round,
  scale_y_increment
)

Arguments

chart_id
Datawrapper chart id.

y
String. Name of variable mapped to the y-axis.

scale_y_max
Integer. The maximum value of the y-axis range. Defaults to the highest value in the variable.

scale_y_min_rule
Minimum value of the y-axis range, either as an integer or a string. Defaults to "min", which takes the lowest value in the variable. "truncated" leaves some one-third space of the plot area empty between the lowest value in the variable and the start of the scale.

scale_y_round
Number of decimal places for all values in the y-axis. Negative values round to a power of ten.

scale_y_increment
Increment of the y-axis breaks.

See also

Examples

if (FALSE) {
jp_scale_y(
  chart_id = "4BcD3",
  y = "lifeExp",
  scale_y_min_rule = 0,
  scale_y_round = -2, # Rounds to the nearest hundred
  scale_y_increment = 20
)
}